+++ src/main/java/io/smilo/api/block/data/transaction/TransactionDTO.java
@@ -17,6 +17,7 @@
 
 package io.smilo.api.block.data.transaction;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -25,11 +26,11 @@ public class TransactionDTO {
     private static final int VERSION = 1;
 
     private String assetId;
-    private Long inputAmount;
+    private BigInteger inputAmount;
     private List<TransactionOutputDTO> transactionOutputs;
     private Long timestamp;
     private String inputAddress;
-    private Long fee;
+    private BigInteger fee;
     private String signatureData;
     private Long signatureIndex;
     private String dataHash;
@@ -59,11 +60,11 @@ public class TransactionDTO {
         this.assetId = assetId;
     }
 
-    public Long getInputAmount() {
+    public BigInteger getInputAmount() {
         return inputAmount;
     }
 
-    public void setInputAmount(Long inputAmount) {
+    public void setInputAmount(BigInteger inputAmount) {
         this.inputAmount = inputAmount;
     }
 
@@ -91,11 +92,11 @@ public class TransactionDTO {
         this.inputAddress = inputAddress;
     }
 
-    public Long getFee() {
+    public BigInteger getFee() {
         return fee;
     }
 
-    public void setFee(Long fee) {
+    public void setFee(BigInteger fee) {
         this.fee = fee;
     }
